+++ packages/ze-design-website/gatsby-browser.jsx
@@ -2,7 +2,7 @@
 import { GatsbyBrowser } from 'gatsby';
 import React from 'react';
 import { WrapElement } from './gatsby-wrap-element';
-import Providers from './src/context/providers';
+import Providers from './src/providers/providers';
 
 /**
  * Gatsby Browser APIs