+++ packages/ze-design-website/gatsby-wrap-element.tsx
@@ -1,6 +1,3 @@
-/**
- * Import fonts
- */
 import { GatsbyCommonPageContext } from '@newrade/core-gatsby-config';
 import { LazyLayoutDesignSystem } from '@newrade/core-gatsby-ui/src/layout/design-system-lazy.layout';
 import { LazyLayoutDocs } from '@newrade/core-gatsby-ui/src/layout/docs-lazy.layout';
@@ -8,12 +5,12 @@ import '@newrade/ze-design-system/lib/assets/fonts/inter-webfonts/stylesheet.css
 import '@newrade/ze-design-system/lib/assets/fonts/wulkan-display-webfonts/stylesheet.css';
 import { PageProps, WrapPageElementBrowserArgs } from 'gatsby';
 import React from 'react';
-import ProvidersDocs from './src/context/providers-docs';
-import ProvidersSite from './src/context/providers-site';
 import './src/design-system/theme.css';
 import { themeClass } from './src/design-system/theme.css';
 import { cssTheme, light, theme } from './src/design-system/theme.treat';
 import { Layout } from './src/layout/layout';
+import ProvidersDocs from './src/providers/providers-docs';
+import ProvidersSite from './src/providers/providers-site';
 
 type Props = PageProps<{}, GatsbyCommonPageContext>;